Fixed: Incorrect output: only the last season is written

+++ CHAD Model/Model/ClimateModule/Climate.cs
@@ -38,14 +38,14 @@ namespace CHAD.Model.ClimateModule
                 var temperature = Gaussian(seasonNumber,
                     climateForecast.TempMean,
                     climateForecast.TempSD,
-                    (double)_parameters.ClimateChangeTempMean,
-                    (double)_parameters.ClimateChangeTempSD);
+                    _parameters.ClimateChangeTempMean,
+                    _parameters.ClimateChangeTempSD);
 
                 var precipitation = Gaussian(seasonNumber,
                     climateForecast.PrecipMean,
                     climateForecast.PrecipSD,
-                    (double)_parameters.ClimateChangePrecipMean,
-                    (double)_parameters.ClimateChangePrecipSD);
+                    _parameters.ClimateChangePrecipMean,
+                    _parameters.ClimateChangePrecipSD);
 
                 _dailyClimates.Add(new DailyClimate(climateForecast.Day, temperature, precipitation));
             }
@@ -80,7 +80,7 @@ namespace CHAD.Model.ClimateModule
             var x1 = 1 - random.NextDouble();
             var x2 = 1 - random.NextDouble();
             var y1 = Math.Sqrt(-2.0 * Math.Log(x1)) * Math.Cos(2.0 * Math.PI * x2);
-            return (double) Math.Round(y1 * standardDeviation * Math.Pow(changeDeviation, seasonNumber) + mean * Math.Pow(changeMean, seasonNumber), 2);
+            return Math.Round(y1 * standardDeviation * Math.Pow(changeDeviation, seasonNumber) + mean * Math.Pow(changeMean, seasonNumber), 2);
         }
 
         #endregion
